Move JSON parse / stringify operations into runtime
